pyproject.toml
[project]
name = "revlm"
version = "0.1.0"
description = "VLM-based Re-Identification with Appearance Alignment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.13.0",
    "diffusers",
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.92",
    "peft>=0.18.1",
    "pillow>=12.1.1",
    "safetensors>=0.7.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "ultralytics>=8.4.21",
    "yacs>=0.1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: real model integration tests (requires model weights)",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
